+++ zalando-cloud-aws-autoconfigure/pom.xml
@@ -6,12 +6,12 @@
 	<modelVersion>4.0.0</modelVersion>
 	
 	<parent>
-		<groupId>com.zalando.awspring.cloud</groupId>
+		<groupId>org.zalando.awspring.cloud</groupId>
 		<artifactId>zalando-cloud-aws</artifactId>
 		<version>3.1.1-SNAPSHOT</version>
 	</parent>
 	
-	<groupId>com.zalando.awspring.cloud</groupId>
+	<groupId>org.zalando.awspring.cloud</groupId>
 	<artifactId>zalando-cloud-aws-autoconfigure</artifactId>
 	<version>3.1.1-SNAPSHOT</version>
 	<packaging>jar</packaging>
@@ -47,7 +47,7 @@
 		</dependency>
 		
 		<dependency>
-			<groupId>com.zalando.awspring.cloud</groupId>
+			<groupId>org.zalando.awspring.cloud</groupId>
 			<artifactId>zalando-cloud-aws-kms</artifactId>
 			<version>${project.version}</version>
 			<optional>true</optional>
