+++ zalando-cloud-aws-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -3,5 +3,5 @@ org.springframework.cloud.bootstrap.BootstrapConfiguration=\
 io.awspring.cloud.autoconfigure.core.AwsAutoConfiguration,\
 io.awspring.cloud.autoconfigure.core.CredentialsProviderAutoConfiguration,\
 io.awspring.cloud.autoconfigure.core.RegionProviderAutoConfiguration,\
-com.zalando.awsspring.cloud.autoconfigure.kms.KmsAutoConfiguration,\
-com.zalando.awsspring.cloud.bootstrap.encrypt.KmsEncryptConfiguration
+org.zalando.awsspring.cloud.autoconfigure.kms.KmsAutoConfiguration,\
+org.zalando.awsspring.cloud.bootstrap.encrypt.KmsEncryptConfiguration
