+++ zalando-cloud-aws-autoconfigure/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-com.zalando.awsspring.cloud.autoconfigure.kms.KmsAutoConfiguration
+org.zalando.awsspring.cloud.autoconfigure.kms.KmsAutoConfiguration
