+++ zalando-cloud-aws-kms/pom.xml
@@ -6,12 +6,12 @@
 	<modelVersion>4.0.0</modelVersion>
 	
 	<parent>
-		<groupId>com.zalando.awspring.cloud</groupId>
+		<groupId>org.zalando.awspring.cloud</groupId>
 		<artifactId>zalando-cloud-aws</artifactId>
 		<version>3.1.1-SNAPSHOT</version>
 	</parent>
 	
-	<groupId>com.zalando.awspring.cloud</groupId>
+	<groupId>org.zalando.awspring.cloud</groupId>
 	<artifactId>zalando-cloud-aws-kms</artifactId>
 	<version>3.1.1-SNAPSHOT</version>
 	<packaging>jar</packaging>
