--- zalando-cloud-aws-kms/src/main/java/com/zalando/awsspring/cloud/bootstrap/encrypt/OutputMode.java
@@ -1,6 +1,0 @@
-package com.zalando.awsspring.cloud.bootstrap.encrypt;
-
-public enum OutputMode {
-
-	PLAIN, BASE64;
-}
