+++ zalando-cloud-aws-kms/src/main/java/org/zalando/awsspring/cloud/bootstrap/encrypt/OutputMode.java
@@ -1,0 +1,6 @@
+package org.zalando.awsspring.cloud.bootstrap.encrypt;
+
+public enum OutputMode {
+
+	PLAIN, BASE64;
+}
