+++ zalando-cloud-aws-samples/zalando-cloud-aws-kms-sample/pom.xml
@@ -12,7 +12,7 @@
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	
-	<groupId>com.zalando.awspring.cloud</groupId>
+	<groupId>org.zalando.awspring.cloud</groupId>
 	<artifactId>zalando-cloud-aws-kms-sample</artifactId>
 	<version>3.1.1-SNAPSHOT</version>
 	<packaging>jar</packaging>
@@ -53,12 +53,12 @@
 		</dependency>
 		
 		<dependency>
-			<groupId>com.zalando.awspring.cloud</groupId>
+			<groupId>org.zalando.awspring.cloud</groupId>
 			<artifactId>zalando-cloud-aws-kms</artifactId>
 			<version>${project.version}</version>
 		</dependency>
 		<dependency>
-			<groupId>com.zalando.awspring.cloud</groupId>
+			<groupId>org.zalando.awspring.cloud</groupId>
 			<artifactId>zalando-cloud-aws-autoconfigure</artifactId>
 			<version>${project.version}</version>
 		</dependency>
