+++ zalando-cloud-aws-samples/zalando-cloud-aws-kms-sample/src/main/java/org/zalando/awsspring/samples/bootstrap/BootstrapApplication.java
@@ -1,0 +1,12 @@
+package org.zalando.awsspring.samples.bootstrap;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BootstrapApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BootstrapApplication.class, args);
+	}
+}
