+++ zalando-cloud-aws-starters/zalando-cloud-aws-starter-kms/pom.xml
@@ -6,13 +6,14 @@
 
 	<parent>
 		<artifactId>zalando-cloud-aws</artifactId>
-		<groupId>com.zalando.awspring.cloud</groupId>
+		<groupId>org.zalando.awspring.cloud</groupId>
 		<version>3.1.1-SNAPSHOT</version>
 		<relativePath>../../pom.xml</relativePath>
 	</parent>
 
 	<modelVersion>4.0.0</modelVersion>
 
+    <groupId>org.zalando.awspring.cloud</groupId>
 	<artifactId>zalando-cloud-aws-starter-kms</artifactId>
 	<name>Zalando Cloud AWS KMS Starter</name>
 	<description>Zalando Cloud AWS Key Management Service Starter</description>
@@ -23,7 +24,7 @@
 			<artifactId>spring-cloud-aws-starter</artifactId>
 		</dependency>
 		<dependency>
-			<groupId>com.zalando.awspring.cloud</groupId>
+			<groupId>org.zalando.awspring.cloud</groupId>
 			<artifactId>zalando-cloud-aws-kms</artifactId>
 			<version>${project.version}</version>
 		</dependency>
